+++ user-management/src/main/java/in/cdac/university/usm/repository/UserRepository.java
@@ -15,7 +15,7 @@ public interface UserRepository extends JpaRepository<UmmtUserMst, UmmtUserMstPK
 
     List<UmmtUserMst> findAllByUniversityMstUnumUnivIdAndGnumIsvalidOrderByGstrUserName(Integer universityId, Integer isValid);
 
-    List<UmmtUserMst> findAllByGnumUseridInAndGnumIsvalidIn(List<Integer> userIds, List<Integer> isValid);
+    List<UmmtUserMst> findAllByGnumUseridInAndGnumIsvalidIn(List<Long> userIds, List<Integer> isValid);
 
     Optional<UmmtUserMst> findByGstrUserNameIgnoreCase(String username);
 
@@ -24,21 +24,21 @@ public interface UserRepository extends JpaRepository<UmmtUserMst, UmmtUserMstPK
             "a.gdtLstmodDate = now(), " +
             "a.gnumLstmodBy = ?1 " +
             "where a.gnumUserid in (?2) and a.gnumIsvalid in (1, 2) ")
-    Integer deleteUsers(Integer modifyByUserId, List<Integer> userIds);
+    Integer deleteUsers(Long modifyByUserId, List<Long> userIds);
 
     @Modifying(clearAutomatically = true)
     @Query("update UmmtUserMst u set u.gnumIsvalid = (select coalesce(max(a.gnumIsvalid), 2) + 1 " +
             "from UmmtUserMst a where a.gnumUserid = u.gnumUserid and a.gnumIsvalid > 2) " +
             "where u.gnumUserid in (?1) and u.gnumIsvalid in (1, 2) ")
-    Integer createLog(Integer userId);
+    Integer createLog(Long userId);
 
-    Optional<UmmtUserMst> findByGstrUserNameIgnoreCaseAndGnumUseridNot(String username, Integer userId);
+    Optional<UmmtUserMst> findByGstrUserNameIgnoreCaseAndGnumUseridNot(String username, Long userId);
 
     @Query("select coalesce(max(a.gnumIsvalid), 2) + 1 from UmmtUserMst a where a.gnumUserid = ?1 and a.gnumIsvalid > 2")
-    Integer getIsValidLogId(Integer userId);
+    Integer getIsValidLogId(Long userId);
 
     @Query("select max(a.gnumUserid) + 1 from UmmtUserMst a")
-    Integer generateUserId();
+    Long generateUserId();
 
     List<UmmtUserMst> findByGnumUserCatIdAndGnumIsvalid(Integer gnumUserCatId, Integer gnumIsvalid);
