+++ user-management/src/main/java/in/cdac/university/usm/repository/UserRoleRepository.java
@@ -11,11 +11,11 @@ import java.util.List;
 @Repository
 public interface UserRoleRepository extends JpaRepository<UmmtUserRoleMst, Long> {
 
-    List<UmmtUserRoleMst> findByGnumUserIdAndGblIsvalid(Integer gnumUserId, Integer gblIsvalid);
+    List<UmmtUserRoleMst> findByGnumUserIdAndGblIsvalid(Long gnumUserId, Integer gblIsvalid);
 
     @Query("select c from UmmtUserRoleMst c " +
             "where c.gnumUserId = :gnumUserId " +
             "and c.gblIsvalid = :gblIsvalid " +
             "and (c.gnumIsDefault = :gnumIsDefault or c.gnumRoleId = :gnumRoleId)")
-    List<UmmtUserRoleMst> findMappedRoles(@Param("gnumUserId") Integer gnumUserId, @Param("gblIsvalid") Integer gblIsvalid, @Param("gnumIsDefault") Integer gnumIsDefault, @Param("gnumRoleId") Integer gnumRoleId);
+    List<UmmtUserRoleMst> findMappedRoles(@Param("gnumUserId") Long gnumUserId, @Param("gblIsvalid") Integer gblIsvalid, @Param("gnumIsDefault") Integer gnumIsDefault, @Param("gnumRoleId") Integer gnumRoleId);
 }
